# chapter-05/project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "List Items",
                "responses": {
                    "200": {
                        "description": "Success: List all Items",
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create Item",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Item",
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get Item by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display thhe specified Item by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Item",
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update Item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Item by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Item updated",
                        "schema": {
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete Item",
                "description": "Delete the specified Item by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Item": {
            "required": [
                "type",
                "name",
                "company"
            ],
            "properties": {
                "type": {
                    "description": "Item Type",
                    "type": "string",
                    "example": "Exhaust"
                },
                "name": {
                    "description": "Item name",
                    "type": "string",
                    "example": "2 into 1 Exhaust"
                },
                "company": {
                    "description": "Produced by: some company",
                    "type": "string",
                    "example": "Vance and Hines"
                }
            }
        },
...
    }
}